+++ src/cider/nrepl.clj
@@ -18,6 +18,8 @@
    [nrepl.middleware.session :refer [session]]
    [nrepl.misc :as misc :refer [with-session-classloader]]
    [nrepl.server :as nrepl-server]
+   [orchard.info]
+   [orchard.namespace]
    [orchard.java]))
 
 ;; Perform the underlying dynamic `require`s asap, and also not within a separate thread
@@ -26,24 +28,39 @@
 @@orchard.java/parser-next-available?
 @analyzer/spec-abbrev
 
-;; Warm up this cache, drastically improving the completion and info UX performance for first hits.
-;; (This was our behavior for many years,
-;; then had to be disabled for test suite reasons in Orchard 0.15.0 to 0.17.0 / cider-nrepl 0.38.0 to 0.41.0, and now it's restored again)
-;; Note that this can only be done cider-nrepl side, unlike before when it was done in Orchard itself.
+(defn warmup-orchard-caches!
+  "Warms up `orchard.java/cache`, drastically improving the completion and info UX performance for first hits.
+  (This was our behavior for many years,
+  then had to be disabled for test suite reasons in Orchard 0.15.0 to 0.17.0 / cider-nrepl 0.38.0 to 0.41.0, and now it's restored again)
+  Note that this can only be done cider-nrepl side, unlike before when it was done in Orchard itself."
+  []
+  ;; 1.- Warmup the overall cache for core Java and Clojure stuff
+  @orchard.java/cache-initializer
+  ;; 2.- Also cache members involved in a typical exception (which includes stuff from nrepl, clojure.repl, etc)
+  @analyzer/ns-common-prefix
+  (try
+    (/ 2 0) ;; throw an exception for its analysis
+    (catch Exception e
+      ;; This performs a many `orchard.info/info*` calls (one per stacktrace frame).
+      ;; Without analyzing one exception in advance, first-time usages of Haystack will be noticeably slow (#828) - prevent that:
+      (analyzer/analyze e)))
+  ;; 3.- Also cache classes that are `:import`ed throughout the project.
+  ;;     The class list is obtanied through `ns` form analysis,
+  ;;      so that we don't depend on whether the namespaces have been loaded yet:
+  (doseq [ns-form (vals (orchard.namespace/project-ns-forms))
+          class-sym (orchard.namespace/ns-form-imports ns-form)
+          :when (try
+                  (Class/forName (str class-sym)
+                                 false ;; Don't initialize this class, avoiding side-effects
+                                 (.getContextClassLoader (Thread/currentThread)))
+                  (catch Throwable _))]
+    (orchard.info/info 'user class-sym))
+  ;; 4.- Leave an indicator that can help up assess the size of a cache in a future:
+  (-> orchard.java/cache .keySet .size))
+
 (def initializer
   (future
-    ;; 1.- Warmup the overall cache for core Java and Clojure stuff
-    @orchard.java/cache-initializer
-    ;; 2.- Also cache members involved in a typical exception (which includes stuff from nrepl, clojure.repl, etc)
-    (try
-      @analyzer/ns-common-prefix
-      (/ 2 0) ;; throw an exception for its analysis
-      (catch Exception e
-        ;; This performs a many `orchard.info/info*` calls (one per stacktrace frame).
-        ;; Without analyzing one exception in advance, first-time usages of Haystack will be noticeably slow (#828) - prevent that:
-        (analyzer/analyze e)))
-    ;; Leave an indicator that can help up assess the size of a cache in a future:
-    (-> orchard.java/cache keys count)))
+    (warmup-orchard-caches!)))
 
 ;;; Functionality for deferred middleware loading
 ;;